Split .core to .core.platform, .core.scenario; adjust imports
