+++ ixmp/backend/__init__.py
@@ -1,5 +1,5 @@
 from enum import IntFlag
-from typing import Dict, Type
+from typing import Dict, List, Type, Union
 
 #: Lists of field names for tuples returned by Backend API methods.
 #:
@@ -40,6 +40,9 @@ FIELDS = {
         "VALUE",
     ),
 }
+
+#: Partial dimensions for “IAMC format”.
+IAMC_IDX: List[Union[str, int]] = ["model", "scenario", "region", "variable", "unit"]
 
 
 #: Mapping from names to available backends. To register additional backends,
