+++ ixmp/tests/core/test_timeseries.py
@@ -7,7 +7,7 @@ from numpy import testing as npt
 from pandas.testing import assert_frame_equal
 
 from ixmp import Scenario, TimeSeries
-from ixmp.core import IAMC_IDX
+from ixmp.backend import IAMC_IDX
 from ixmp.testing import DATA, models
 
 # string columns for timeseries checks
